# docs/swagger.yaml
definitions:
  serializer.Article:
    properties:
      ArticleContent:
        type: string
      ArticleId:
        type: integer
      ArticleStat:
        type: integer
      ArticleTags:
        type: string
      ArticleTime:
        type: string
      ArticleTitle:
        type: string
      AuthorId:
        type: integer
      AuthorName:
        type: string
    type: object
  serializer.Comment:
    properties:
      commentID:
        type: integer
      content:
        type: string
      revUserName:
        type: string
      stat:
        type: integer
      subComment:
        type: boolean
      time:
        type: string
      userID:
        type: integer
      userName:
        type: string
    type: object
  serializer.User:
    properties:
      Briefly:
        type: string
      FansNum:
        type: integer
      FollowerNum:
        type: integer
      UserName:
        type: string
    type: object
host: localhost
info:
  contact:
    email: 773532732@qq.com
    name: 陈浩
    url: https://github.com/chen102
  description: 完成了博客的基本功能
  title: 博客系统后端
  version: "1.0"
paths:
  /api/v0/article/add:
    post:
      description: 用户增加文章
      parameters:
      - description: 文章标题名，最大20
        in: body
        name: ArticleTitle
        required: true
        schema:
          type: string
      - description: 文章正文
        in: body
        name: ArticleContent
        required: true
        schema:
          type: string
      - description: 文章标签
        in: body
        name: Tags
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 发表成功
          schema:
            type: string
      summary: 增加文章
  /api/v0/article/comment:
    post:
      description: 显示文章评论列表
      parameters:
      - description: 文章id
        in: body
        name: ArticleId
        required: true
        schema:
          type: integer
      - description: 评论id，若为空即为对文章评论
        in: body
        name: Rev
        schema:
          type: integer
      - description: 列表偏移量
        in: body
        name: Offset
        schema:
          type: integer
      - description: 列表一页请求的个数
        in: body
        name: Count
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.Comment'
            type: array
      summary: 评论列表
  /api/v0/article/delete:
    post:
      description: 用户删除指定的文章
      parameters:
      - description: 删除id
        in: body
        name: DeleteId
        required: true
        schema:
          type: integer
      - description: 文章id,若是删除评论，这个字段不能为空
        in: body
        name: ArticleId
        schema:
          type: integer
      - description: 若为空，删文章，为true删评论
        in: body
        name: Type
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            type: string
      summary: 删除文章/评论
  /api/v0/article/list:
    post:
      description: 显示用户文章列表
      parameters:
      - description: 用户id，若为空，即为当前登录用户
        in: body
        name: AuthorId
        schema:
          type: integer
      - description: 排序：若为空按时间排序，若为true按发布时间排序
        in: body
        name: Rank
        schema:
          type: boolean
      - description: 列表偏移量
        in: body
        name: Offset
        schema:
          type: integer
      - description: 列表一页请求的个数
        in: body
        name: Count
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.Article'
      summary: 用户文章列表
  /api/v0/article/show:
    get:
      description: 显示文章详情
      parameters:
      - description: 文章id
        in: body
        name: ArticleId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.Article'
            type: array
      summary: 查看文章
  /api/v0/article/stat:
    post:
      description: 用户给文章或者评论点赞
      parameters:
      - description: 点赞id
        in: body
        name: StatId
        required: true
        schema:
          type: integer
      - description: 点赞类型:若为空，对文章点赞，为true点赞评论
        in: body
        name: StatType
        schema:
          type: boolean
      - description: 取消点赞:若为空，点赞，为true取消点赞
        in: body
        name: CancelStat
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
      summary: 点赞
  /api/v0/follower:
    post:
      description: 用户关注另一个用户
      parameters:
      - description: 用户id
        in: body
        name: UserId
        required: true
        schema:
          type: integer
      - description: 若为true对该用户取关
        in: body
        name: CancelFollower
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 关注成功
          schema:
            type: string
      summary: 关注用户
  /api/v0/follower/dynamic:
    post:
      description: 显示关注用户的动态
      parameters:
      - description: 列表偏移量
        in: body
        name: Offset
        schema:
          type: integer
      - description: 列表一页请求的个数
        in: body
        name: Count
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.Article'
            type: array
      summary: 用户关注动态
  /api/v0/follower/list:
    post:
      description: 显示用户关注的用户列表信息
      parameters:
      - description: 用户id，若为空，即为当前登录用户
        in: body
        name: UserId
        schema:
          type: integer
      - description: '1:只能是1或为空，空:关注列表，1:'
        in: body
        name: Type
        schema:
          type: integer
      - description: 列表偏移量
        in: body
        name: Offset
        schema:
          type: integer
      - description: 列表一页请求的个数
        in: body
        name: Count
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.User'
            type: array
      summary: 用户关注/粉丝列表
  /api/v0/user/like:
    post:
      description: 显示用户点赞的文章列表
      parameters:
      - description: 用户id，若为空，即为当前登录用户
        in: body
        name: AuthorId
        schema:
          type: integer
      - description: 列表偏移量
        in: body
        name: Offset
        schema:
          type: integer
      - description: 列表一页请求的个数
        in: body
        name: Count
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.Article'
            type: array
      summary: 用户喜欢文章列表
  /api/v0/user/login:
    post:
      description: 用户输入账号密码进行登录
      parameters:
      - description: 账号,长度大于7，小于20
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密码,长度大于7，小于20
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            type: string
      summary: 登录用户
  /api/v0/user/register:
    post:
      description: 用户输入账号密码等信息进行注册
      parameters:
      - description: 用户名,长度大于2，小于20
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 账号,长度大于7，小于20
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密码,长度大于7，小于20
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 重复密码
        in: body
        name: reppassword
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            type: string
      summary: 注册用户
swagger: "2.0"